Can't seem to add print() in RandTransform.before_call
A test that passed the first time does not pass after
a print statement is added to `RandomTransform.before_call`.

Steps to reproduce the observed behaviour:
() Shutdown '09_vision_augment.ipynb'.
() Open '09_vision_augment.ipynb'.
() Execute notebook up until `In [14]`.
() Go to the definition of `RandTransform.before_call` (`In [6]`).
() Add `print(self.do)` at the end of this function. Run cell.
() Go back down to `In [14]`.
() Either re-execute this cell, or copy the code to a new cell and execute it.